smart-agriculture-iot/backend/ml_requirements.txt
# Smart Agriculture IoT System - ML Requirements
# Plant Disease Detection Dependencies

# Core ML libraries
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0

# Web framework for API
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0

# Image processing
scikit-image==0.21.0
matplotlib==3.7.1

# Data handling
pandas==2.0.3

# Optional: For production deployment
gunicorn==21.2.0

# Development tools
python-dotenv==1.0.0
